% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_All_Files.R
\name{clean_all_files}
\alias{clean_all_files}
\title{Clean all eddy covariance data files in a directory}
\usage{
clean_all_files(directory = "~/Eddy Covariance Data/",
  time_zone = "Asia/Manila")
}
\arguments{
\item{directory}{The directory where the eddy covariance data files are
stored that you wish to use to calculate ET values}

\item{time_zone}{The time zone for which the data is native. Defaults to Asia/
Manila for IRRI HQ. Should not be changed unless you know what you are doing
and are outside of IRRI HQ}
}
\description{
This function automates filtering and gap filling of eddy-
covariance data by filling gaps using functions from zoo and uses
a Hampel filter from the R package \code{\link{pracma}} to filter the time
series and remove outliers. Use this function to clean a series of eddy
covariance data.
}
\examples{
\dontrun{
clean_all_files(directory = "~/Eddy Covariance Data/")
}

}
