% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_One_Day.R
\name{clean_one_day}
\alias{clean_one_day}
\title{Clean One Day (24 hour) Eddy Covariance Data File}
\usage{
clean_one_day(directory = "~/Eddy Covariance Data/",
  time_zone = "Asia/Manila")
}
\arguments{
\item{directory}{The directory where the eddy covariance data files are
stored that you wish to use to calculate ET values}

\item{time_zone}{The time zone for which the data is native. Defaults to Asia/
Manila for IRRI HQ. Should not be changed unless you know what you are doing
and are outside of IRRI HQ}
}
\description{
This function automates filtering and gap filling of eddy-
covariance data from the IRRI Ecological Intensification Platform for
calculating evapotranspiration for irrigation scheduling. This function
checks a specified directory for the most recent data file from the eddy
covariance tower and the most recent previous file. It then takes the two
files and runs the same operations (gap filling and a Hampel filter) and
then generates an estimate of evapotranspiration for the previous 24 hour
time-period. IRRI uses this to then schedule pivot irrigation in the
ecological intensification platform.
}
\examples{
\dontrun{
clean_one_day(directory = "~/Eddy Covariance Data/")
}

}

